chore: enable tests and merge them into one class
Bean counts affects service scoped bean and the counter should not be used for service scoped beans.
Other tests reset counts and if the are executed before service bean tests the counter has incorrect value.
Personal counter is used to count service scoped beans and then all other service tests are merged into one class.
